eom-proj/1-class-data/requirements.txt
Flask-Bootstrap4
flask-wtf
